Revert "Implement Python-Based Serverless API for BST Inference"
