+++ README.md
@@ -165,9 +165,37 @@ Notebooks and configuration for deploying trained models to Google Cloud VertexA
 - VertexAI endpoint deployment
 - Terraform infrastructure setup
 
+### BST Model Export and Optimization
+Tools for exporting and optimizing BST (Badminton Stroke-type Transformer) models for cloud inference:
 
-### BST Model Export and Serverless API
-See [bst_export/README.md](bst_export/README.md) for all details on exporting BST models and deploying the serverless API.
+#### Quick Start
+```bash
+# Export BST model to TorchScript and ONNX formats
+python export_bst_model.py --model_type BST_CG_AP --weights_path weights/bst_model.pt
+
+# Test exported models for cloud deployment
+python cloud_deployment_example.py --model_path weights/exported/bst_cg_ap_seq100_scripted.pt
+
+# Run basic functionality tests
+python test_export_basic.py
+```
+
+#### Makefile Commands
+```bash
+make export-bst        # Export BST model with default settings
+make export-bst-all    # Export to all formats with benchmarking
+make test-export       # Test export functionality
+make test-cloud        # Test cloud deployment simulation
+```
+
+#### Key Features
+- **Multiple Model Variants**: BST_0, BST, BST_CG, BST_AP, BST_CG_AP
+- **Dual Export Formats**: TorchScript (.pt) and ONNX (.onnx)
+- **Cloud Optimization**: Memory and inference speed optimization
+- **Deployment Examples**: Google Cloud Functions, AWS Lambda, ONNX Runtime
+- **Performance Benchmarking**: Inference speed and memory usage testing
+
+See [docs/BST_MODEL_EXPORT_GUIDE.md](docs/BST_MODEL_EXPORT_GUIDE.md) for comprehensive documentation.
 
 ### Experiments (experiments/)
 Experimental notebooks for testing new approaches:
@@ -199,31 +227,6 @@ from modeling.feature_engineering import preprocess_pose_sequence
 # Predict shot type from pose sequence
 pose_sequence = preprocess_pose_sequence(pose_data)
 prediction = predict_shot(model, pose_sequence)
-```
-
-### Using the Serverless API
-
-```python
-import requests
-import numpy as np
-
-# Create sample badminton pose data
-sample_data = {
-    "JnB": np.random.randn(1, 100, 2, 72).tolist(),  # Pose features
-    "shuttle": np.random.randn(1, 100, 2).tolist(),   # Shuttle trajectory  
-    "pos": np.random.randn(1, 100, 2, 2).tolist(),    # Player positions
-    "video_len": [100]                                # Video length
-}
-
-# Make prediction request
-response = requests.post(
-    "http://localhost:8000/predict",  # Or your Cloud Functions URL
-    json=sample_data
-)
-
-result = response.json()
-print(f"Top shot prediction: {result['top_predictions']['indices'][0][0]}")
-print(f"Confidence: {result['top_predictions']['probabilities'][0][0]:.3f}")
 ```
 
 ## Troubleshooting
