--- bst_export_serverless/.gitignore
@@ -1,1 +1,0 @@
-.venv_bst_export
