--- bst_export_serverless/requirements_serverless.txt
@@ -1,26 +1,0 @@
-# Serverless API Requirements
-# Minimal dependencies for BST model inference
-
-# Core API framework
-fastapi>=0.100.0
-uvicorn>=0.23.0
-pydantic>=2.0.0
-
-# Essential ML dependencies
-numpy>=1.24.0
-
-# Model inference backends (one or both)
-torch>=2.0.0  # For TorchScript models
-onnxruntime>=1.15.0  # For ONNX models
-
-# Google Cloud Functions support
-functions-framework>=3.0.0
-
-# Optional: For AWS Lambda deployment
-# aws-lambda-powertools[pydantic]
-
-# For making HTTP requests (e.g., downloading models)
-requests==2.32.5
-
-# For softmax function in serverless_api.py
-scipy==1.16.2
